separated event control panel from mainwindow
